refactor: Extract expression parser
